(main): cast i to u_long to avoid signed/unsigned comparison

+++ tests/Reactor_Timer_Test.cpp
@@ -4,7 +4,7 @@
 //
 // = LIBRARY
 //    tests
-// 
+//
 // = FILENAME
 //    Reactor_Timer_Test.cpp
 //
@@ -16,7 +16,7 @@
 //
 // = AUTHOR
 //    Prashant Jain and Douglas C. Schmidt
-// 
+//
 // ============================================================================
 
 #include "test_config.h"
@@ -37,12 +37,12 @@ public:
       ACE_ASSERT (current_count == count);
 
       ACE_DEBUG ((LM_DEBUG, "%d: Timer #%d timed out at %d!\n",
-		  count, current_count, tv.sec ()));
+                  count, current_count, tv.sec ()));
 
       count += (1 + odd);
 
       if (current_count == ACE_MAX_TIMERS - 1)
-	done = 1;
+        done = 1;
 
       return 0;
     }
@@ -58,40 +58,40 @@ main (int, char *[])
   Time_Handler rt[ACE_MAX_TIMERS];
   int t_id[ACE_MAX_TIMERS];
   long i;
-  
-  for (i = 0; i < ACE_MAX_TIMERS; i++)
-    t_id[i] = reactor.schedule_timer (&rt[i], 
-				      (const void *) i, 
-				      ACE_Time_Value (2 * i + 1));
+
+  for (i = 0; (u_long) i < ACE_MAX_TIMERS; i++)
+    t_id[i] = reactor.schedule_timer (&rt[i],
+                                      (const void *) i,
+                                      ACE_Time_Value (2 * i + 1));
   while (!done)
     reactor.handle_events ();
-  
+
   done = 0;
   count = 0;
-  
+
   // Now try multiple timers for ONE event handler (should produce the
   // same result).
-  for (i = 0; i < ACE_MAX_TIMERS; i++)
+  for (i = 0; (u_long) i < ACE_MAX_TIMERS; i++)
     t_id[i] = reactor.schedule_timer (&rt[0],
-				      (const void *) i, 
-				      ACE_Time_Value (2 * i + 1));
+                                      (const void *) i,
+                                      ACE_Time_Value (2 * i + 1));
 
   while (!done)
     reactor.handle_events ();
-  
+
   done = 0;
   count = 1;
   odd = 1;
 
-  for (i = 0; i < ACE_MAX_TIMERS; i++)
+  for (i = 0; (u_long) i < ACE_MAX_TIMERS; i++)
     {
-      t_id[i] = reactor.schedule_timer (&rt[0], 
-					(const void *) i, 
-					ACE_Time_Value (2 * i + 1));
+      t_id[i] = reactor.schedule_timer (&rt[0],
+                                        (const void *) i,
+                                        ACE_Time_Value (2 * i + 1));
 
       // Cancel even numbered timers.
       if (ACE_EVEN (i))
-	reactor.cancel_timer (t_id[i]);
+        reactor.cancel_timer (t_id[i]);
     }
 
   while (!done)
